requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
pymongo==3.12.3
pytz==2024.1
sqlparse>=0.3.1
typing-extensions==4.12.2